Optimization Analysis

Successful type inference and optimization is key to high-performing Julia programs. But as mentioned in the performance tips, there are some chances where Julia can not infer the types of your program very well and can not optimize it well accordingly.

While there are many possibilities of "type-instabilities", like usage of non-constant global variable most notably, probably the most tricky one would be "captured variable" – Julia can not really well infer the type of variable that is observed and modified by both inner function and enclosing one. And such type instabilities can lead to various optimization failures. One of the most common barrier to the performance is known as "runtime dispatch", which happens when a matching method can't be resolved by the compiler due to the lack of type information and it is looked up at runtime instead. Since runtime dispatch is caused by poor type information, it often indicates the compiler could not do other optimizations including inlining and scalar replacements of aggregates.

In order to avoid such problems, we usually insect output of code_typed or its family, and check if there is anywhere type is not well inferred and optimization was not successful. But the problem is that one needs to have enough knowledge about the inference and optimization in order to interpret the output. Another problem is that they can only present the "final" output of the inference and optimization, and we can not inspect the entire call graph and may miss to find where a problem actually happened and how the type-instability has been propagated. There is a nice package called Cthulhu.jl, which allows us to look at the outputs of code_typed by descending into a call tree, recursively and interactively. The workflow with Cthulhu is much more efficient and powerful, but still, it requires much familiarity with Julia compiler and it tends to be tedious.

So, why not automate it ? JET implements such an analyzer that investigates optimized representation of your program and automatically detects anywhere the compiler failed in optimization. Especially, it can find where Julia creates captured variables, where runtime dispatch will happen, and where Julia gives up the optimization work due to unresolvable recursive function call.

Quick Start

JET exports @report_opt, which analyzes the entire call graph of a given generic function call, and then reports detected performance pitfalls:

julia> using JET

As a first example, let's see how we can find and fix runtime dispatches using JET:

julia> n = rand(Int); # non-constant global variable
julia> make_vals(n) = n ≥ 0 ? (zero(n):n) : (n:zero(n));
julia> function sumup(f)
           # this function uses the non-constant global variable `n` here
           # and it makes every succeeding operations type-unstable
           vals = make_vals(n)
           s = zero(eltype(vals))
           for v in vals
               s += f(v)
           end
           return s
       end;
julia> @report_opt sumup(sin) # runtime dispatches will be reported═════ 7 possible errors found ═════
┌ @ REPL[3]:4 Main.make_vals(%1)
│ runtime dispatch detected: Main.make_vals(%1::Any)
└─────────────
┌ @ REPL[3]:5 Main.eltype(%2)
│ runtime dispatch detected: Main.eltype(%2::Any)
└─────────────
┌ @ REPL[3]:5 Main.zero(%3)
│ runtime dispatch detected: Main.zero(%3::Any)
└─────────────
┌ @ REPL[3]:6 Base.iterate(%2)
│ runtime dispatch detected: Base.iterate(%2::Any)
└─────────────
┌ @ REPL[3]:7 f(%11)
│ runtime dispatch detected: f::typeof(sin)(%11::Any)
└─────────────
┌ @ REPL[3]:7 Main.+(%10, %13)
│ runtime dispatch detected: Main.+(%10::Any, %13::Any)
└─────────────
┌ @ REPL[3]:8 Base.iterate(%2, %12)
│ runtime dispatch detected: Base.iterate(%2::Any, %12::Any)
└─────────────

JET's analysis result will be dynamically updated when we update fucntion definition[1], and we can "hot-fix" the runtime dispatches within the same running Julia session like this:

julia> # we can pass parameters as a function argument instead, and then everything will be type-stable
       function sumup(f, n)
           vals = make_vals(n)
           s = zero(eltype(vals))
           for v in vals
               # NOTE here we may get union type like `s::Union{Int,Float64}`,
               # but Julia can optimize away such small unions (thus no runtime dispatch)
               s += f(v)
           end
           return s
       end;
julia> @report_opt sumup(sin, rand(Int)) # now runtime dispatch free !No errors !

@report_opt can also report existence of captured variables, which are really better to be eliminated within performance-sensitive context:

julia> # the examples below are all adapted from https://docs.julialang.org/en/v1/manual/performance-tips/#man-performance-captured
       function abmult(r::Int)
           if r < 0
               r = -r
           end
           # the closure assigned to `f` make the variable `r` captured
           f = x -> x * r
           return f
       end;
julia> @report_opt abmult(42)═════ 3 possible errors found ═════
┌ @ REPL[1]:2 r = Core.Box(:(_7::Int64))
│ captured variable `r` detected
└─────────────
┌ @ REPL[1]:2 Main.<(%7, 0)
│ runtime dispatch detected: Main.<(%7::Any, 0)
└─────────────
┌ @ REPL[1]:3 Main.-(%14)
│ runtime dispatch detected: Main.-(%14::Any)
└─────────────
julia> function abmult(r0::Int)
           # we can improve the type stability of the variable `r` like this,
           # but it is still captured
           r::Int = r0
           if r < 0
               r = -r
           end
           f = x -> x * r
           return f
       end;
julia> @report_opt abmult(42)═════ 1 possible error found ═════
┌ @ REPL[3]:4 r = Core.Box()
│ captured variable `r` detected
└─────────────
julia> function abmult(r::Int)
           if r < 0
               r = -r
           end
           # we can try to eliminate the capturing
           # and now this function would be most high-performing
           f = let r = r
               x -> x * r
           end
           return f
       end;
julia> @report_opt abmult(42)No errors !

With the frame_filter configuration, we can easily limit the scope of analysis to specific module context:

julia> # problem: when ∑1/n exceeds `x` ?
       function compute(x)
           r = 1
           s = 0.0
           n = 1
           @time while r < x
               s += 1/n
               if s ≥ r
                   # `println` call is full of runtime dispatches for good reasons
                   # and we're not interested in type-instabilities within this call
                   # since we know it's only called few times
                   println("round $r/$x has been finished")
                   r += 1
               end
               n += 1
           end
           return n, s
       endcompute (generic function with 1 method)
julia> @report_opt compute(30) # bunch of reports will be reported from the `println` call═════ 23 possible errors found ═════
┌ @ REPL[1]:11 Main.println(Base.string("round ", r, "/", x, " has been finished"))
│┌ @ coreio.jl:4 Base.println(Core.tuple(Core.typeassert(Base.stdout, Base.IO)), xs...)
││┌ @ strings/io.jl:73 Base.print(Core.tuple(io), xs, Core.tuple("\n")...)
│││┌ @ strings/io.jl:43 Base.lock(io)
││││┌ @ stream.jl:283 Base.lock(Base.getproperty(s, :lock))
│││││┌ @ lock.jl:100 Base.wait(Base.getproperty(rl, :cond_wait))
││││││┌ @ condition.jl:112  = Base.wait()
│││││││┌ @ task.jl:824 result = Base.try_yieldto(Base.ensure_rescheduled)
││││││││┌ @ task.jl:761 Base.getproperty(%7, :result)
│││││││││ runtime dispatch detected: Base.getproperty(%7::Task, :result::Symbol)
││││││││└───────────────
││││││││┌ @ task.jl:762 Base.setproperty!(%7, :result, Base.nothing)
│││││││││ runtime dispatch detected: Base.setproperty!(%7::Task, :result::Symbol, Base.nothing)
││││││││└───────────────
││││││││┌ @ task.jl:763 Base.setproperty!(%7, :_isexception, false)
│││││││││ runtime dispatch detected: Base.setproperty!(%7::Task, :_isexception::Symbol, false)
││││││││└───────────────
││││││┌ @ condition.jl:114 Base.list_deletefirst!(Base.getproperty(ct, :queue), ct)
│││││││┌ @ linked_list.jl:145 Base.isequal(Base.getproperty(h, :value), val)
││││││││┌ @ gcutils.jl:4 Base.isequal(%1, v)
│││││││││ runtime dispatch detected: Base.isequal(%1::Any, v::Task)
││││││││└────────────────
││││││┌ @ condition.jl:114 Base.list_deletefirst!(%56, %50)
│││││││ runtime dispatch detected: Base.list_deletefirst!(%56::Any, %50::Task)
││││││└────────────────────
│││┌ @ strings/io.jl:49 Base.unlock(io)
││││┌ @ show.jl:335 Base.unlock(Base.getproperty(io, :io))
│││││┌ @ stream.jl:284 Base.unlock(Base.getproperty(s, :lock))
││││││┌ @ lock.jl:132 Base.notify(Base.getproperty(rl, :cond_wait))
│││││││┌ @ condition.jl:130 #self#(c, Base.nothing)
││││││││┌ @ condition.jl:130 Base.#notify#564(true, false, #self#, c, arg)
│││││││││┌ @ condition.jl:130 Base.notify(c, arg, all, error)
││││││││││┌ @ condition.jl:136 Core.kwfunc(Base.schedule)(Core.apply_type(Core.NamedTuple, (:error,))(Core.tuple(error)), Base.schedule, t, arg)
│││││││││││┌ @ task.jl:684 Base.#schedule#585(error, _3, t, arg)
││││││││││││┌ @ task.jl:686 %10(%11, t)
│││││││││││││ runtime dispatch detected: %10::typeof(Base.list_deletefirst!)(%11::Any, t::Task)
││││││││││││└───────────────
│┌ @ coreio.jl:4 Base.println(%3, %4)
││ runtime dispatch detected: Base.println(%3::IO, %4::String)
│└───────────────
┌ @ timing.jl:214 Base.time_print(elapsedtime, Base.getproperty(diff, :allocd), Base.getproperty(diff, :total_time), Base.gc_alloc_count(diff), compile_elapsedtime, true)
│┌ @ timing.jl:120 str = Base.sprint(#866)
││┌ @ strings/io.jl:106 Base.#sprint#430(Core.tuple(Base.nothing, 0, #self#, f), args...)
│││┌ @ strings/io.jl:112 f(Core.tuple(s), args...)
││││┌ @ timing.jl:123 Base.!=(%32, 0)
│││││ runtime dispatch detected: Base.!=(%32::Any, 0)
││││└─────────────────
││││┌ @ timing.jl:125 Base.!=(%65, 0)
│││││ runtime dispatch detected: Base.!=(%65::Any, 0)
││││└─────────────────
││││┌ @ timing.jl:126 Base.prettyprint_getunits(%73, %75, 1000)
│││││ runtime dispatch detected: Base.prettyprint_getunits(%73::Any, %75::Int64, 1000)
││││└─────────────────
││││┌ @ timing.jl:127 Base.==(%80, 1)
│││││ runtime dispatch detected: Base.==(%80::Any, 1)
││││└─────────────────
││││┌ @ timing.jl:128 Base.Int(%88)
│││││ runtime dispatch detected: Base.Int(%88::Any)
││││└─────────────────
││││┌ @ timing.jl:128 Base.getindex(Base._cnt_units, %80)
│││││ runtime dispatch detected: Base.getindex(Base._cnt_units, %80::Any)
││││└─────────────────
││││┌ @ timing.jl:128 Base.==(%96, 1)
│││││ runtime dispatch detected: Base.==(%96::Any, 1)
││││└─────────────────
││││┌ @ timing.jl:128 Base.print(io, %89, %90, %101)
│││││ runtime dispatch detected: Base.print(io::IOBuffer, %89::Any, %90::Any, %101::String)
││││└─────────────────
││││┌ @ timing.jl:130 Base.Float64(%110)
│││││ runtime dispatch detected: Base.Float64(%110::Any)
││││└─────────────────
││││┌ @ timing.jl:130 %104(%111, 2)
│││││ runtime dispatch detected: %104::typeof(Base.Ryu.writefixed)(%111::Any, 2)
││││└─────────────────
││││┌ @ timing.jl:130 Base.getindex(Base._cnt_units, %80)
│││││ runtime dispatch detected: Base.getindex(Base._cnt_units, %80::Any)
││││└─────────────────
││││┌ @ timing.jl:130 Base.print(io, %112, %113, " allocations: ")
│││││ runtime dispatch detected: Base.print(io::IOBuffer, %112::String, %113::Any, " allocations: ")
││││└─────────────────
││││┌ @ timing.jl:135 Base.!=(%138, 0)
│││││ runtime dispatch detected: Base.!=(%138::Any, 0)
││││└─────────────────
││││┌ @ timing.jl:141 Base.!=(%173, 0)
│││││ runtime dispatch detected: Base.!=(%173::Any, 0)
││││└─────────────────
│┌ @ timing.jl:148 Base.print(str)
││┌ @ coreio.jl:3 Base.print(%3, %4)
│││ runtime dispatch detected: Base.print(%3::IO, %4::String)
││└───────────────
│┌ @ timing.jl:119 allocs = Core.Box(:(_11::Int64))
││ captured variable `allocs` detected
│└─────────────────
julia> this_module_filter(sv) = sv.mod === @__MODULE__;
julia> @report_opt frame_filter=this_module_filter compute(30) # focus on what we wrote, and no error should be reportedNo errors !

There is also function_filter, which can ignore specific function call.

@test_opt can be used to assert that a given function call is free from the performance pitfalls. It is fully integrated with Test standard library's unit-testing infrastructure, and we can use it as like other Test macros e.g. @test:

julia> @test_opt sumup(cos)JET-test failed at REPL[1]:1
  Expression: #= REPL[1]:1 =# JET.@test_call analyzer = JET.OptAnalyzer sumup(cos)
  ═════ 7 possible errors found ═════
  ┌ @ REPL[3]:4 Main.make_vals(%1)
  │ runtime dispatch detected: Main.make_vals(%1::Any)
  └─────────────
  ┌ @ REPL[3]:5 Main.eltype(%2)
  │ runtime dispatch detected: Main.eltype(%2::Any)
  └─────────────
  ┌ @ REPL[3]:5 Main.zero(%3)
  │ runtime dispatch detected: Main.zero(%3::Any)
  └─────────────
  ┌ @ REPL[3]:6 Base.iterate(%2)
  │ runtime dispatch detected: Base.iterate(%2::Any)
  └─────────────
  ┌ @ REPL[3]:7 f(%11)
  │ runtime dispatch detected: f::typeof(cos)(%11::Any)
  └─────────────
  ┌ @ REPL[3]:7 Main.+(%10, %13)
  │ runtime dispatch detected: Main.+(%10::Any, %13::Any)
  └─────────────
  ┌ @ REPL[3]:8 Base.iterate(%2, %12)
  │ runtime dispatch detected: Base.iterate(%2::Any, %12::Any)
  └─────────────
  
ERROR: There was an error during testing
julia> @test_opt frame_filter=this_module_filter compute(30)Test Passed
  Expression: #= REPL[2]:1 =# JET.@test_call analyzer = JET.OptAnalyzer frame_filter = this_module_filter compute(30)
julia> using Test
julia> @testset "check type-stabilities" begin
           @test_opt sumup(cos) # should fail
       
           n = rand(Int)
           @test_opt sumup(cos, n) # should pass
       
           @test_opt frame_filter=this_module_filter compute(30) # should pass
       
           @test_opt broken=true compute(30) # should pass with the "broken" annotation
       endcheck type-stabilities: JET-test failed at REPL[4]:2
  Expression: #= REPL[4]:2 =# JET.@test_call analyzer = JET.OptAnalyzer sumup(cos)
  ═════ 7 possible errors found ═════
  ┌ @ REPL[3]:4 Main.make_vals(%1)
  │ runtime dispatch detected: Main.make_vals(%1::Any)
  └─────────────
  ┌ @ REPL[3]:5 Main.eltype(%2)
  │ runtime dispatch detected: Main.eltype(%2::Any)
  └─────────────
  ┌ @ REPL[3]:5 Main.zero(%3)
  │ runtime dispatch detected: Main.zero(%3::Any)
  └─────────────
  ┌ @ REPL[3]:6 Base.iterate(%2)
  │ runtime dispatch detected: Base.iterate(%2::Any)
  └─────────────
  ┌ @ REPL[3]:7 f(%11)
  │ runtime dispatch detected: f::typeof(cos)(%11::Any)
  └─────────────
  ┌ @ REPL[3]:7 Main.+(%10, %13)
  │ runtime dispatch detected: Main.+(%10::Any, %13::Any)
  └─────────────
  ┌ @ REPL[3]:8 Base.iterate(%2, %12)
  │ runtime dispatch detected: Base.iterate(%2::Any, %12::Any)
  └─────────────
  
Test Summary:          | Pass  Fail  Broken  Total
check type-stabilities |    2     1       1      4
ERROR: Some tests did not pass: 2 passed, 1 failed, 0 errored, 1 broken.

Entry Points

These macros/functions are the entries of dispatch analysis:

JET.@report_opt — Macro

@report_opt [jetconfigs...] f(args...)

Evaluates the arguments to the function call, determines its types, and then calls report_opt on the resulting expression. As with @code_typed and its family, any of JET configurations or optimization analysis specific configurations can be given as the optional arguments like this:

# reports `rand(::Type{Bool})` with `unoptimize_throw_blocks` configuration turned on
julia> @report_opt unoptimize_throw_blocks=true rand(Bool)

JET.report_opt — Function

report_opt(f, types = Tuple{}; jetconfigs...) -> JETCallResult
report_opt(tt::Type{<:Tuple}; jetconfigs...) -> JETCallResult

Analyzes the generic function call with the given type signature with the optimization analyzer, which collects optimization failures and runtime dispatches involved within the call stack.

JET.@test_opt — Macro

@test_opt [jetconfigs...] [broken=false] [skip=false] f(args...)

Tests the generic function call f(args...) is free from runtime dispatch. Returns a Pass result if it is, a Fail result if if contains any location where runtime dispatch or optimization failure happens, or an Error result if this macro encounters an unexpected error. When the test Fails, abstract call stack to each problem location will also be printed to stdout.

julia> @test_opt sincos(10)
Test Passed
  Expression: #= none:1 =# JET.@test_opt sincos(10)

As with @report_opt, any of JET configurations or optimization analysis specific configurations can be given as the optional arguments like this:

julia> function f(n)
            r = sincos(n)
            println(r) # `println` is full of runtime dispatches, but we can ignore the corresponding reports from `Base` by explicit frame filter
            return r
       end;
julia> this_module_filter(x) = x.mod === @__MODULE__;

julia> @test_opt frame_filter=this_module_filter f(10)
Test Passed
  Expression: #= none:1 =# JET.@test_opt frame_filter = this_module_filter f(10)

@test_opt is fully integrated with Test standard library's unit-testing infrastructure. It means, the result of @test_opt will be included in the final @testset summary, it supports skip and broken annotations as @test macro does, etc.

julia> using Test

julia> f(params) = sin(params.value); # type-stable
julia> params = (; value = 10);       # non-constant global variable
julia> g() = sin(params.value);       # very type-instable

julia> @testset "check optimizations" begin
           @test_opt f((; value = 10)) # pass
           @test_opt g()               # fail
           @test_opt broken=true g()   # annotated as broken, thus still "pass"
       end
check optimizations: Dispatch Test Failed at none:3
  Expression: #= none:3 =# JET.@test_opt g()
  ═════ 2 possible errors found ═════
  ┌ @ none:1 Base.getproperty(%1, :value)
  │ runtime dispatch detected: Base.getproperty(%1::Any, :value::Symbol)
  └──────────
  ┌ @ none:1 Main.sin(%2)
  │ runtime dispatch detected: Main.sin(%2::Any)
  └──────────

Test Summary:       | Pass  Fail  Broken  Total
check optimizations |    1     1       1      3
ERROR: Some tests did not pass: 1 passed, 1 failed, 0 errored, 1 broken.

JET.test_opt — Function

test_opt(f, types = Tuple{}; broken::Bool = false, skip::Bool = false, jetconfigs...)
test_opt(tt::Type{<:Tuple}; broken::Bool = false, skip::Bool = false, jetconfigs...)

Tests the generic function call with the given type signature is free from runtime dispatch. Except that it takes a type signature rather than a call expression, this function works in the same way as @test_opt.

Configurations

JET.OptAnalyzer — Type

Every entry point of optimization analysis can accept any of JET configurations as well as the following additional configurations that are specific to optimization analysis.

frame_filter = x::Union{Core.Compiler.InferenceState, Core.Compiler.OptimizationState}->true:
A predicate which takes InfernceState or OptimizationState and returns false to skip analysis on the frame.

# only checks code within the current module:
julia> mymodule_filter(x) = x.mod === @__MODULE__;

julia> @test_opt frame_filter=mymodule_filter f(args...)
...

function_filter = @nospecialize(ft)->true:
A predicate which takes a function type and returns false to skip analysis on the call.

# ignores `Core.Compiler.widenconst` calls (since it's designed to be runtime-dispatched):
julia> myfunction_filter(@nospecialize(ft)) = ft !== typeof(Core.Compiler.widenconst)

julia> @test_opt function_filter=myfunction_filter f(args...)
...

skip_nonconcrete_calls::Bool = true:
Julia's runtime dispatch is "powerful" because it can always compile code with concrete runtime arguments so that a "kernel" function runs very effectively even if it's called from a type-instable call site. This means, we (really) often accept that some parts of our code are not inferred statically, and rather we want to just rely on information that is only available at runtime. To model this programming style, the optimization analyzer does NOT report any optimization failures or runtime dispatches detected within non-concrete calls under the default configuration. We can turn off this skip_nonconcrete_calls configuration to get type-instabilities within non-concrete calls.

# the following examples are adapted from https://docs.julialang.org/en/v1/manual/performance-tips/#kernel-functions
julia> function fill_twos!(a)
           for i = eachindex(a)
               a[i] = 2
           end
       end;

julia> function strange_twos(n)
           a = Vector{rand(Bool) ? Int64 : Float64}(undef, n)
           fill_twos!(a)
           return a
       end;

# by default, only type-instabilities within concrete call (i.e. `strange_twos(3)`) are reported
# and those within non-concrete calls (`fill_twos!(a)`) are not reported
julia> @report_opt strange_twos(3)
═════ 2 possible errors found ═════
┌ @ REPL[2]:2 %45(Main.undef, n)
│ runtime dispatch detected: %45::Type{Vector{_A}} where _A(Main.undef, n::Int64)
└─────────────
┌ @ REPL[2]:3 Main.fill_twos!(%46)
│ runtime dispatch detected: Main.fill_twos!(%46::Vector)
└─────────────
Vector (alias for Array{_A, 1} where _A)

# we can get reports from non-concrete calls with `skip_nonconcrete_calls=false`
julia> @report_opt skip_nonconcrete_calls=false strange_twos(3)
═════ 4 possible errors found ═════
┌ @ REPL[2]:3 Main.fill_twos!(a)
│┌ @ REPL[1]:3 Base.setindex!(a, 2, %14)
││ runtime dispatch detected: Base.setindex!(a::Vector, 2, %14::Int64)
│└─────────────
│┌ @ REPL[1]:3 Base.setindex!(a, 2, i)
││┌ @ array.jl:877 Base.convert(_, x)
│││ runtime dispatch detected: Base.convert(_::Any, x::Int64)
││└────────────────
┌ @ REPL[2]:2 %45(Main.undef, n)
│ runtime dispatch detected: %45::Type{Vector{_A}} where _A(Main.undef, n::Int64)
└─────────────
┌ @ REPL[2]:3 Main.fill_twos!(%46)
│ runtime dispatch detected: Main.fill_twos!(%46::Vector)
└─────────────
Vector (alias for Array{_A, 1} where _A)

skip_unoptimized_throw_blocks::Bool = true:
By default, Julia's native compilation pipeline intentionally disables inference (and so succeeding optimizations too) on "throw blocks", which are code blocks that will eventually lead to throw calls, in order to ease the compilation latency problem, a.k.a. "first-time-to-plot". Accordingly, the optimization analyzer also ignores any performance pitfalls detected within those blocks since we usually don't mind if code involved with error handling isn't optimized. If skip_unoptimized_throw_blocks is set to false, it doesn't ignore them and will report type instabilities detected within "throw blocks".

See also https://github.com/JuliaLang/julia/pull/35982.

# by default, unoptimized "throw blocks" are not analyzed
julia> @test_opt sin(10)
Test Passed
  Expression: #= none:1 =# JET.@test_opt sin(10)

# we can turn on the analysis on unoptimized "throw blocks" with `skip_unoptimized_throw_blocks=false`
julia> @test_opt skip_unoptimized_throw_blocks=false sin(10)
JET-test failed at none:1
  Expression: #= REPL[6]:1 =# JET.@test_call analyzer = JET.OptAnalyzer skip_unoptimized_throw_blocks = false sin(10)
  ═════ 1 possible error found ═════
  ┌ @ math.jl:1221 Base.Math.sin(xf)
  │┌ @ special/trig.jl:39 Base.Math.sin_domain_error(x)
  ││┌ @ special/trig.jl:28 Base.Math.DomainError(x, "sin(x) is only defined for finite x.")
  │││ runtime dispatch detected: Base.Math.DomainError(x::Float64, "sin(x) is only defined for finite x.")
  ││└──────────────────────

ERROR: There was an error during testing

# we can also turns off the heuristic itself
julia> @test_opt unoptimize_throw_blocks=false skip_unoptimized_throw_blocks=false sin(10)
Test Passed
  Expression: #= REPL[7]:1 =# JET.@test_call analyzer = JET.OptAnalyzer unoptimize_throw_blocks = false skip_unoptimized_throw_blocks = false sin(10)

1Technically, it's fully integrated with Julia's method invalidation system.