How to Use JET.jl

Tip

Hold on tight and fasten your seatbelt while JET is analyzing your code for the first time. Once the caches get accumulated, subsequent analysis will run at ✈ speed.

Top-level Entry Points

JET can analyze your "top-level" code. This means you can just give your Julia file or code to JET and get error reports. report_file, report_and_watch_file, report_package and report_text are the main entry points for that.

JET will analyze your code "half-statically" – JET will selectively interpret "top-level definitions" (like a function definition) and try to simulate Julia's top-level code execution, while it tries to avoid executing any other parts of code like function calls but analyzes them by abstract interpretation instead (and this is a part where JET statically analyzes your code). If you're interested in how JET selects "top-level definitions", please see JET.virtual_process.

Warning

Because JET will actually interpret "top-level definitions" in your code, it certainly runs your code. So we should note that JET can cause some side effects from your code; for example JET will try to expand all the macros used in your code, and so the side effects involved with macro expansions will also happen in JET's analysis process.

JET.report_file — Function

report_file(filename::AbstractString;
            jetconfigs...) -> JETToplevelResult

Analyzes filename and returns JETToplevelResult.

This function will look for .JET.toml configuration file in the directory of filename, and search up the file tree until any .JET.toml is (or isn't) found. When found, the configurations specified in the file will be applied. See Configuration File for more details.

Tip

When you want to analyze your package, but any file actually using it isn't available, the analyze_from_definitions option can be useful (see ToplevelConfig's analyze_from_definitions option).
For example, JET can analyze JET itself like below:

# from the root directory of JET.jl
julia> report_file("src/JET.jl";
                   analyze_from_definitions = true)

See also: report_package

Note

This function will enable the toplevel_logger configuration by default with the default logging level. You can still explicitly specify and configure it:

report_file(args...;
            toplevel_logger = nothing, # suppress toplevel logger
            jetconfigs...) # other configurations

See Logging Configurations for more details.

JET.report_and_watch_file — Function

report_and_watch_file(filename::AbstractString;
                      jetconfigs...)

Watches filename and keeps re-triggering analysis with report_file on code update. JET will try to analyze all the included files reachable from filename, and it will re-trigger analysis if there is code update detected in any of the included files.

This function internally uses Revise.jl to track code updates. Revise also offers possibilities to track changes in files that are not directly analyzed by JET, or even changes in Base files. See Watch Configurations for more details.

See also: report_file

JET.report_package — Function

report_package(package::Union{AbstractString,Module};
               jetconfigs...) -> JETToplevelResult

Analyzes package in the same way as report_file with the special default configurations, which are especially tuned for package analysis (see below for details). package can be either a module or a string. In the latter case it must be the name of a package in your current environment.

This function configures analysis with the following configurations:

analyze_from_definitions = true: allows JET to enter analysis without top-level call sites; this is useful for package analysis since a package itself usually has only definitions but not usages (i.e. call sites)

concretization_patterns = [:(x_ = y_), :(const x_ = y_)]: concretizes every global variable instantiations; concretizations are generally preferred for successful analysis as far as they're cheap, and global variable instantiations that occur in a package definition are usually very cheap

See ToplevelConfig for more details.

report_package([io::IO = stdout];
               jetconfigs...) -> res::ReportResult

Like above but analyzes the package of the current project.

See also: report_file

JET.report_text — Function

report_text(text::AbstractString,
            filename::AbstractString = "top-level";
            jetconfigs...) -> JETToplevelResult

Analyzes text and returns JETToplevelResult.

Interactive Entry Points

There are utilities for checking JET analysis in a running Julia session like REPL or such.

JET.report_call — Function

report_call(f, types = Tuple{};
            analyzer::Type{<:AbstractAnalyzer} = JETAnalyzer,
            jetconfigs...) -> JETCallResult
report_call(tt::Type{<:Tuple};
            analyzer::Type{<:AbstractAnalyzer} = JETAnalyzer,
            jetconfigs...) -> JETCallResult

Analyzes the generic function call with the given type signature with analyzer. And finally returns the analysis result as JETCallResult.

JET.@report_call — Macro

@report_call [jetconfigs...] f(args...)

Evaluates the arguments to the function call, determines its types, and then calls report_call on the resulting expression. As with @code_typed and its family, any of JET configurations can be given as the optional arguments like this:

# reports `rand(::Type{Bool})` with `aggressive_constant_propagation` configuration turned off
julia> @report_call aggressive_constant_propagation=false rand(Bool)

Test Integration

JET offers entries that are integrated with Test standard library's unit-testing infrastructure. It can be used in your test suite to assert your program is free from errors that JET can detect. Currently, it only supports interactive entries.

JET.@test_call — Macro

@test_call [jetconfigs...] [broken=false] [skip=false] f(args...)

Runs @report_call jetconfigs... f(args...) and tests that the generic function call f(args...) is free from problems that @report_call can detect. If executed inside @testset, returns a Pass result if it is, a Fail result if it contains any error points detected, or an Error result if this macro encounters an unexpected error. When the test Fails, abstract call stack to each problem location will also be printed to stdout.

julia> @test_call sincos(10)
Test Passed
  Expression: #= none:1 =# JET.@test_call sincos(10)

As with @report_call, any of JET configurations or analyzer specific configurations can be given as the optional arguments jetconfigs... like this:

julia> cond = false

julia> function f(n)
            if cond           # `cond` is untyped, and will be reported by the sound analysis pass, while JET's default analysis pass will ignore it
                return sin(n)
            else
                return cos(n)
            end
       end;

julia> @test_call f(10)
Test Passed
  Expression: #= none:1 =# JET.@test_call f(10)

julia> @test_call mode=:sound f(10)
JET-test failed at none:1
  Expression: #= none:1 =# JET.@test_call mode = :sound f(10)
  ═════ 1 possible error found ═════
  ┌ @ none:2 goto %4 if not Main.cond
  │ non-boolean (Any) used in boolean context: goto %4 if not Main.cond
  └──────────

ERROR: There was an error during testing

@test_call is fully integrated with Test standard library's unit-testing infrastructure. It means, the result of @test_call will be included in the final @testset summary, it supports skip and broken annotations as like @test and its family:

julia> using JET, Test

julia> f(ref) = isa(ref[], Number) ? sin(ref[]) : nothing;      # Julia can't propagate the type constraint `ref[]::Number` to `sin(ref[])`, JET will report `NoMethodError`

julia> g(ref) = (x = ref[]; isa(x, Number) ? sin(x) : nothing); # we can make it type-stable if we extract `ref[]` into a local variable `x`

julia> @testset "check errors" begin
           ref = Ref{Union{Nothing,Int}}(0)
           @test_call f(ref)             # fail
           @test_call g(ref)             # fail
           @test_call broken=true f(ref) # annotated as broken, thus still "pass"
       end
check errors: JET-test failed at none:3
  Expression: #= none:3 =# JET.@test_call f(ref)
  ═════ 1 possible error found ═════
  ┌ @ none:1 Main.sin(Base.getindex(ref))
  │ for 1 of union split cases, no matching method found for call signatures (Tuple{typeof(sin), Nothing})): Main.sin(Base.getindex(ref::Base.RefValue{Union{Nothing, Int64}})::Union{Nothing, Int64})
  └──────────

Test Summary: | Pass  Fail  Broken  Total
check errors  |    1     1       1      3
ERROR: Some tests did not pass: 1 passed, 1 failed, 0 errored, 1 broken.

JET.test_call — Function

test_call(f, types = Tuple{}; broken::Bool = false, skip::Bool = false, jetconfigs...)
test_call(tt::Type{<:Tuple}; broken::Bool = false, skip::Bool = false, jetconfigs...)

Runs report_call(f, types; jetconfigs... and tests that the generic function call f(args...) is free from problems that report_call can detect. Except that it takes a type signature rather than a call expression, this function works in the same way as @test_call.