Configurations

JET analysis can be flexibly fine-tuned. Any entry point explained in JET's default error analysis and the optimization analysis can accept any of the configuration parameters described below as keyword arguments (or optional parameters for the interactive macros). For example, you can analyze the call of sum("julia") with the annotate_types configuration enabled as like:

@report_call annotate_types=true sum("julia")

or equivalently:

report_call(sum, (String,); annotate_types=true)

Similarly you can analyze a top-level script path/to/file.jl with specifying target_defined_modules configuration as follows:

report_file("path/to/file.jl";
            target_defined_modules = true)

Note

Please ignore the names of documented objects appearing below, like "JET.configured_reports". They are just remnants of documentation internals, and you will never directly interact with them.

Configurations for Analysis Result

JET.configured_reports — Function

Configurations for JET's analysis results. These configurations are always active.

target_modules = nothing
A configuration to filter out reports by specifying module contexts where problems should be reported.

By default (target_modules = nothing), JET reports all detected problems. If specified, a problem is reported if its module context matches any of target_modules settings and hidden otherwise. target_modules should be an iterator of whose element is either of the data types below that match report::InferenceErrorReport's context module as follows:

m::Module or JET.LastFrameModule(m::Module): matches if the module context of report's innermost stack frame is m

JET.AnyFrameModule(m::Module): matches if module context of any of report's stack frame is m

user-type T: matches according to user-definition overload match_module(::T, report::InferenceErrorReport)

ignored_modules = nothing
A configuration to filter out reports by specifying module contexts where problems should be ignored.

By default (ignored_modules = nothing), JET reports all detected problems. If specified, a problem is hidden if its module context matches any of ignored_modules settings and reported otherwise. ignored_modules should be an iterator of whose element is either of the data types below that match report::InferenceErrorReport's context module as follows:

m::Module or JET.LastFrameModule(m::Module): matches if the module context of report's innermost stack frame is m

JET.AnyFrameModule(m::Module): matches if module context of any of report's stack frame is m

user-type T: matches according to user-definition overload match_module(::T, report::InferenceErrorReport)

report_config = nothing
Additional configuration layer to filter out reports with user-specified strategies. By default (report_config = nothing), JET will use the module context based configurations elaborated above and below. If user-type T is given, then JET will report problems based on the logic according to an user-overload configured_reports(::T, reports::Vector{InferenceErrorReport}), and the target_modules and ignored_modules configurations are not really active.

Examples

julia> function foo(a)
           r1 = sum(a)       # => Base: MethodError(+(::Char, ::Char)), MethodError(zero(::Type{Char}))
           r2 = undefsum(a)  # => @__MODULE__: UndefVarError(:undefsum)
           return r1, r2
       end;

# by default, JET will print all the collected reports:
julia> @report_call foo("julia")
═════ 3 possible errors found ═════
┌ @ REPL[1]:2 r1 = sum(a)
│┌ @ reduce.jl:549 Base.:(var"#sum#281")(pairs(NamedTuple()), #self#, a)
││┌ @ reduce.jl:549 sum(identity, a)
│││┌ @ reduce.jl:520 Base.:(var"#sum#280")(pairs(NamedTuple()), #self#, f, a)
││││┌ @ reduce.jl:520 mapreduce(f, Base.add_sum, a)
│││││┌ @ reduce.jl:294 Base.:(var"#mapreduce#277")(pairs(NamedTuple()), #self#, f, op, itr)
││││││┌ @ reduce.jl:294 mapfoldl(f, op, itr)
│││││││┌ @ reduce.jl:162 Base.:(var"#mapfoldl#273")(Base._InitialValue(), #self#, f, op, itr)
││││││││┌ @ reduce.jl:162 Base.mapfoldl_impl(f, op, init, itr)
│││││││││┌ @ reduce.jl:44 Base.foldl_impl(op′, nt, itr′)
││││││││││┌ @ reduce.jl:48 v = Base._foldl_impl(op, nt, itr)
│││││││││││┌ @ reduce.jl:62 v = op(v, y[1])
││││││││││││┌ @ reduce.jl:81 op.rf(acc, x)
│││││││││││││┌ @ reduce.jl:24 x + y
││││││││││││││ no matching method found for `+(::Char, ::Char)`: (x::Char + y::Char)::Union{}
│││││││││││││└────────────────
││││││││││┌ @ reduce.jl:49 Base.reduce_empty_iter(op, itr)
│││││││││││┌ @ reduce.jl:370 Base.reduce_empty_iter(op, itr, Base.IteratorEltype(itr))
││││││││││││┌ @ reduce.jl:371 Base.reduce_empty(op, eltype(itr))
│││││││││││││┌ @ reduce.jl:347 Base.reduce_empty(op.rf, T)
││││││││││││││┌ @ reduce.jl:339 Base.reduce_empty(+, T)
│││││││││││││││┌ @ reduce.jl:330 zero(T)
││││││││││││││││ no matching method found for `zero(::Type{Char})`: zero(T::Type{Char})::Union{}
│││││││││││││││└─────────────────
┌ @ REPL[1]:3 r2 = undefsum(a)
│ `undefsum` is not defined
└─────────────

# with `target_modules=(@__MODULE__,)`, JET will only report the problems detected within the `@__MODULE__` module:
julia> @report_call target_modules=(@__MODULE__,) foo("julia")
════ 1 possible error found ═════
┌ @ REPL[1]:3 r2 = undefsum(a)
│ `undefsum` is not defined
└─────────────

# with `ignored_modules=(Base,)`, JET will ignore the errors detected within the `Base` module:
julia> @report_call ignored_modules=(Base,) foo("julia")
════ 1 possible error found ═════
┌ @ REPL[1]:3 r2 = undefsum(a)
│ `undefsum` is not defined
└─────────────

Configurations for Top-level Analysis

JET.ToplevelConfig — Type

Configurations for top-level analysis. These configurations will be active for all the top-level entries explained in the top-level analysis entry points section.

context::Module = Main
The module context in which the top-level execution will be simulated.

This configuration can be useful when you just want to analyze a submodule, without starting entire analysis from the root module. For example, we can analyze Base.Math like below:

julia> report_file(JET.fullbasepath("math.jl");
                   context = Base,                  # `Base.Math`'s root module
                   analyze_from_definitions = true, # there're only definitions in `Base`
                   )

Note that this module context will be virtualized by default so that JET can repeat analysis in the same session without having "invalid redefinition of constant ..." error etc. In other word, JET virtualize the module context of context and make sure the original module context isn't polluted by JET.

target_defined_modules::Bool = false
If true, automatically set the target_modules configuration so that JET filters out errors that are reported within modules that JET doesn't analyze directly.

analyze_from_definitions::Bool = false
If true, JET will start analysis using signatures of top-level definitions (e.g. method signatures), after the top-level interpretation has been done (unless no serious top-level error has happened, like errors involved within a macro expansion).

This can be handy when you want to analyze a package, which usually contains only definitions but not their usages (i.e. top-level callsites). With this option, JET can enter analysis just with method or type definitions, and we don't need to pass a file that uses the target package.

Warning

This feature is very experimental at this point, and you may face lots of false positive errors, especially when trying to analyze a big package with lots of dependencies. If a file that contains top-level callsites (e.g. test/runtests.jl) is available, JET analysis using the file is generally preferred, since analysis entered from concrete call sites will produce more accurate results than analysis entered from (maybe not concrete-typed) method signatures.

Also see: report_file, report_and_watch_file

concretization_patterns::Vector{<:Any} = Expr[]
Specifies a customized top-level code concretization strategy.

When analyzing a top-level code, JET first splits the entire code into appropriate units of code (i.e. code blocks), and then iterate a virtual top-level code execution process on each code block in order to simulate Julia's sequential top-level code execution. In virtual code execution, JET will selectively interpret "top-level definitions" (like a function definition) just like Julia's top-level code execution, while it tries to avoid executing any other parts of code like function calls and leaves them to succeeding static analysis by abstract interpretation.

However, currently, JET doesn't track the "inter-code-block" level code dependencies, and so the selective interpretation of top-level definitions can fail if it needs an access to global variables defined in other code blocks that are not actually interpreted (i.e. "concretized") but just left for abstract interpreation (i.e. "abstracted").

For example, the issue happens when your macro accesses to a global variable during its expansion, e.g.:

test/fixtures/concretization_patterns.jl

# JET doesn't conretize this by default, but just analyzes its type
const GLOBAL_CODE_STORE = Dict()

macro with_code_record(a)
    GLOBAL_CODE_STORE[__source__] = a # record the code location in the global store
    esc(a)
end

# here JET will try to actually expand `@with_code_record`,
# but since `GLOBAL_CODE_STORE` didn't get concretized (i.e. instantiated), JET analysis will fail at this point
@with_code_record foo(a) = identity(a)

foo(10) # top-level callsite, abstracted away

To circumvent this issue, JET offers the concretization_patterns::Vector{<:Any} configuration, which allows us to customize JET's top-level code concretization strategy. concretization_patterns specifies the patterns of code that should be concretized. To put in other word, when JET sees a code that matches any of code patterns specified by an user, JET will try to interpret and concretize the code, regardless of whether JET's code selection logic decides to concretize it or not.

JET uses MacroTools.jl's expression pattern match, and we can specify whatever code pattern expected by MacroTools.@capture macro. For example, in order to solve the issue explained above, we can have:

concretization_patterns = [:(const GLOBAL_CODE_STORE = Dict())]

Then GLOBAL_CODE_STORE will just be concretized and so any top-level error won't happen at the macro expansion.

Although configuring concretization_patterns properly could be really tricky, we can effectively debug JET's top-level code concretization plan using the toplevel_logger configuration with the logging level above than 1 ("debug") level. With the toplevel_logger configuration, we can see:

which code is matched with concretization_patterns and forcibly concretized

which code is selected to be concretized or not by JET's code selection logic: where t-annotated statements are concretized while f-annotated statements are abstracted and left abstract interpretation

julia> report_file("test/fixtures/concretization_patterns.jl";
                   concretization_patterns = [:(const GLOBAL_CODE_STORE = Dict())],
                   toplevel_logger = IOContext(stdout, :JET_LOGGER_LEVEL => 1))

[toplevel-debug] virtualized the context of Main (took 0.003 sec)
[toplevel-debug] entered into test/fixtures/concretization_patterns.jl
[toplevel-debug] concretization pattern `const GLOBAL_CODE_STORE = Dict()` matched `const GLOBAL_CODE_STORE = Dict()` at test/fixtures/concretization_patterns.jl:2
[toplevel-debug] concretization plan at test/fixtures/concretization_patterns.jl:4:
1 f 1 ─      $(Expr(:thunk, CodeInfo(
    @ none within `top-level scope`
1 ─     return $(Expr(:method, Symbol("@with_code_record")))
)))
2 t │        $(Expr(:method, Symbol("@with_code_record")))
3 t │   %3 = Core.Typeof(var"@with_code_record")
4 t │   %4 = Core.svec(%3, Core.LineNumberNode, Core.Module, Core.Any)
5 t │   %5 = Core.svec()
6 t │   %6 = Core.svec(%4, %5, $(QuoteNode(:(#= test/fixtures/concretization_patterns.jl:4 =#))))
7 t │        $(Expr(:method, Symbol("@with_code_record"), :(%6), CodeInfo(
    @ test/fixtures/concretization_patterns.jl:5 within `none`
1 ─      $(Expr(:meta, :nospecialize, :(a)))
│        Base.setindex!(GLOBAL_CODE_STORE, a, __source__)
│   @ test/fixtures/concretization_patterns.jl:6 within `none`
│   %3 = esc(a)
└──      return %3
)))
8 f └──      return var"@with_code_record"
[toplevel-debug] concretization plan at test/fixtures/concretization_patterns.jl:11:
1 f 1 ─      $(Expr(:thunk, CodeInfo(
    @ none within `top-level scope`
1 ─     return $(Expr(:method, :foo))
)))
2 t │        $(Expr(:method, :foo))
3 t │   %3 = Core.Typeof(foo)
4 t │   %4 = Core.svec(%3, Core.Any)
5 t │   %5 = Core.svec()
6 t │   %6 = Core.svec(%4, %5, $(QuoteNode(:(#= test/fixtures/concretization_patterns.jl:11 =#))))
7 t │        $(Expr(:method, :foo, :(%6), CodeInfo(
    @ test/fixtures/concretization_patterns.jl:11 within `none`
1 ─ %1 = identity(a)
└──      return %1
)))
8 f └──      return foo
[toplevel-debug] concretization plan at test/fixtures/concretization_patterns.jl:13:
1 f 1 ─ %1 = foo(10)
2 f └──      return %1
[toplevel-debug]  exited from test/fixtures/concretization_patterns.jl (took 0.032 sec)

Also see: the toplevel_logger section below, virtual_process.

toplevel_logger::Union{Nothing,IO} = nothing
If IO object is given, it will track JET's toplevel analysis. Logging level can be specified with :JET_LOGGER_LEVEL IO property. Currently supported logging levels are either of 0 ("info" level, default), 1 ("debug" level).

Examples:

logs into stdout

julia> report_file(filename; toplevel_logger = stdout)

logs into io::IOBuffer with "debug" logger level

julia> report_file(filename; toplevel_logger = IOContext(io, :JET_LOGGER_LEVEL => 1));

virtualize::Bool = true
When true, JET will virtualize the given root module context.

This configuration is supposed to be used only for testing or debugging. See virtualize_module_context for the internal.

Configurations for Abstract Interpretation

JET.JETInferenceParams — Function

Configurations for abstract interpretation performed by JET. These configurations will be active for all the entries.

You can configure any of the keyword parameters that Core.Compiler.InferenceParams or Core.Compiler.OptimizationParams can take, e.g. max_methods::Int = 3, union_splitting::Int = 4. Listed below are selections of those parameters that can have a potent influence on JET analysis.

ipo_constant_propagation::Bool = true
Enables constant propagation in abstract interpretation. It is highly recommended that you keep this configuration true to get reasonable analysis result, because constant propagation can cut off lots of false positive errorenous code paths and thus produce more accurate and useful analysis results.

aggressive_constant_propagation::Bool = true
If true, JET will try to do constant propagation more "aggressively". It can lead to more accurate analysis as explained above, but also it may incur a performance cost. JET by default enables this configuration to get more accurate analysis result.

unoptimize_throw_blocks::Bool = false
Turn this on to skip analysis on code blocks that will eventually lead to a throw call. This configuration improves the analysis performance, but it's better to be turned off to get a "proper" analysis result, just because there may be other errors even in those "throw blocks".

Print Configurations

JET.PrintConfig — Type

Configurations for report printing. The configurations below will be active whenever showing JET's analysis result within REPL.

annotate_types::Bool = false
When set to true, annotates types when printing analyzed call stack.

Examples:

with annotate_types = false (default):
julia> @report_call sum("julia")
═════ 2 possible errors found ═════
┌ @ reduce.jl:549 Base.:(var"#sum#281")(pairs(NamedTuple()), #self#, a)
│┌ @ reduce.jl:549 sum(identity, a)
││┌ @ reduce.jl:520 Base.:(var"#sum#280")(pairs(NamedTuple()), #self#, f, a)
│││┌ @ reduce.jl:520 mapreduce(f, Base.add_sum, a)
││││┌ @ reduce.jl:294 Base.:(var"#mapreduce#277")(pairs(NamedTuple()), #self#, f, op, itr)
│││││┌ @ reduce.jl:294 mapfoldl(f, op, itr)
││││││┌ @ reduce.jl:162 Base.:(var"#mapfoldl#273")(Base._InitialValue(), #self#, f, op, itr)
│││││││┌ @ reduce.jl:162 Base.mapfoldl_impl(f, op, init, itr)
││││││││┌ @ reduce.jl:44 Base.foldl_impl(op′, nt, itr′)
│││││││││┌ @ reduce.jl:48 v = Base._foldl_impl(op, nt, itr)
││││││││││┌ @ reduce.jl:62 v = op(v, y[1])
│││││││││││┌ @ reduce.jl:81 op.rf(acc, x)
││││││││││││┌ @ reduce.jl:24 x + y
│││││││││││││ no matching method found for `+(::Char, ::Char)`: (x::Char + y::Char)::Union{}
││││││││││││└────────────────
│││││││││┌ @ reduce.jl:49 Base.reduce_empty_iter(op, itr)
││││││││││┌ @ reduce.jl:370 Base.reduce_empty_iter(op, itr, Base.IteratorEltype(itr))
│││││││││││┌ @ reduce.jl:371 Base.reduce_empty(op, eltype(itr))
││││││││││││┌ @ reduce.jl:347 Base.reduce_empty(op.rf, T)
│││││││││││││┌ @ reduce.jl:339 Base.reduce_empty(+, T)
││││││││││││││┌ @ reduce.jl:330 zero(T)
│││││││││││││││ no matching method found for `zero(::Type{Char})`: zero(T::Type{Char})::Union{}
││││││││││││││└─────────────────

with annotate_types = true
julia> @report_call annotate_types = true sum("julia")
═════ 2 possible errors found ═════
┌ @ reduce.jl:549 Base.:(var"#sum#281")(pairs(NamedTuple()::NamedTuple{(), Tuple{}})::Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}}, #self#::typeof(sum), a::String)::Union{}
│┌ @ reduce.jl:549 sum(identity, a::String)::Union{}
││┌ @ reduce.jl:520 Base.:(var"#sum#280")(pairs(NamedTuple()::NamedTuple{(), Tuple{}})::Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}}, #self#::typeof(sum), f::typeof(identity), a::String)::Union{}
│││┌ @ reduce.jl:520 mapreduce(f::typeof(identity), Base.add_sum, a::String)::Union{}
││││┌ @ reduce.jl:294 Base.:(var"#mapreduce#277")(pairs(NamedTuple()::NamedTuple{(), Tuple{}})::Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}}, #self#::typeof(mapreduce), f::typeof(identity), op::typeof(Base.add_sum), itr::String)::Union{}
│││││┌ @ reduce.jl:294 mapfoldl(f::typeof(identity), op::typeof(Base.add_sum), itr::String)::Union{}
││││││┌ @ reduce.jl:162 Base.:(var"#mapfoldl#273")(Base._InitialValue()::Base._InitialValue, #self#::typeof(mapfoldl), f::typeof(identity), op::typeof(Base.add_sum), itr::String)::Union{}
│││││││┌ @ reduce.jl:162 Base.mapfoldl_impl(f::typeof(identity), op::typeof(Base.add_sum), init::Base._InitialValue, itr::String)::Union{}
││││││││┌ @ reduce.jl:44 Base.foldl_impl(op′::Union{}, nt::Base._InitialValue, itr′::Union{})::Union{}
│││││││││┌ @ reduce.jl:48 v = Base._foldl_impl(op::Base.BottomRF{typeof(Base.add_sum)}, nt::Base._InitialValue, itr::String)::Union{}
││││││││││┌ @ reduce.jl:62 v = op::Base.BottomRF{typeof(Base.add_sum)}(v::Char, (y::Tuple{Char, Int64})[1]::Char)::Union{}
│││││││││││┌ @ reduce.jl:81 (op::Base.BottomRF{typeof(Base.add_sum)}).rf::typeof(Base.add_sum)(acc::Char, x::Char)::Union{}
││││││││││││┌ @ reduce.jl:24 (x::Char + y::Char)::Union{}
│││││││││││││ no matching method found for `+(::Char, ::Char)`: (x::Char + y::Char)::Union{}
││││││││││││└────────────────
│││││││││┌ @ reduce.jl:49 Base.reduce_empty_iter(op::Base.BottomRF{typeof(Base.add_sum)}, itr::String)::Union{}
││││││││││┌ @ reduce.jl:370 Base.reduce_empty_iter(op::Base.BottomRF{typeof(Base.add_sum)}, itr::String, Base.IteratorEltype(itr::String)::Base.HasEltype)::Union{}
│││││││││││┌ @ reduce.jl:371 Base.reduce_empty(op::Base.BottomRF{typeof(Base.add_sum)}, eltype(itr::String)::Type{Char})::Union{}
││││││││││││┌ @ reduce.jl:347 Base.reduce_empty((op::Base.BottomRF{typeof(Base.add_sum)}).rf::typeof(Base.add_sum), T::Type{Char})::Union{}
│││││││││││││┌ @ reduce.jl:339 Base.reduce_empty(+, T::Type{Char})::Union{}
││││││││││││││┌ @ reduce.jl:330 zero(T::Type{Char})::Union{}
│││││││││││││││ no matching method found for `zero(::Type{Char})`: zero(T::Type{Char})::Union{}
││││││││││││││└─────────────────

Note

JET always annotates types when printing the error point, e.g. in the example above, the error points below are always type-annotated regardless of this configuration:

no matching method found for call signature: Base.zero(_::Type{Char})

no matching method found for call signature: Base.+(x::Char, y::Char)

fullpath::Bool = false
Controls whether or not expand a file path to full path when printing analyzed call stack. Note that paths of Julia's Base files will also be expanded when set to true.

print_toplevel_success::Bool = false
If true, prints a message when there is no toplevel errors found.

print_inference_success::Bool = true
If true, print a message when there is no errors found in abstract interpretation based analysis pass.

Configurations for VSCode Integration

JET.VSCode.VSCodeConfig — Type

Configurations for the VSCode integration. These configurations are active only when used in the integrated Julia REPL.

vscode_console_output::Union{Nothing,IO} = stdout
JET will show analysis result in VSCode's "PROBLEMS" pane and inline annotations. If vscode_console_output::IO is specified, JET will also print the result into the specified output stream in addition to showing the result in the integrated views. When nothing, the result will be only shown in the integrated views.

Watch Configurations

JET.WatchConfig — Type

Configurations for "watch" mode. The configurations will only be active when used with report_and_watch_file.

revise_all::Bool = true
Redirected to Revise.entr's all keyword argument. When set to true, JET will retrigger analysis as soon as code updates are detected in any module tracked by Revise. Currently when encountering import/using statements, JET won't perform analysis, but rather will just load the modules as usual execution (this also means Revise will track those modules). So if you're editing both files analyzed by JET and modules that are used within the files, this configuration should be enabled.

revise_modules = nothing
Redirected to Revise.entr's modules positional argument. If a iterator of Module is given, JET will retrigger analysis whenever code in modules updates.

Tip

This configuration is useful when your're also editing files that are not tracked by Revise, e.g. editing functions defined in Base:

# re-performe analysis when you make a change to `Base`
julia> report_and_watch_file(yourfile; revise_modules = [Base])

Configuration File

JET.parse_config_file — Function

JET.jl offers .prettierrc style configuration file support. This means you can use .JET.toml configuration file to specify any of configurations explained above and share that with others.

When report_file or report_and_watch_file is called, it will look for .JET.toml in the directory of the given file, and search up the file tree until a JET configuration file is (or isn't) found. When found, the configurations specified in the file will be applied.

A configuration file can specify any of JET configurations like:

aggressive_constant_propagation = false # turn off aggressive constant propagation
... # other configurations

Note that the following configurations should be string(s) of valid Julia code:

context: string of Julia code, which can be parsed and evaluated into Module

concretization_patterns: vector of string of Julia code, which can be parsed into a Julia expression pattern expected by MacroTools.@capture macro.

toplevel_logger: string of Julia code, which can be parsed and evaluated into Union{IO,Nothing}

E.g. the configurations below are equivalent:

configurations via keyword arguments
report_file(somefile;
            concretization_patterns = [:(const GLOBAL_CODE_STORE = x_)],
            toplevel_logger = IOContext(open("toplevel.txt", "w"), :JET_LOGGER_LEVEL => 1))

configurations via a configuration file
# supposed to concretize `const GLOBAL_CODE_STORE = Dict()` in test/fixtures/concretization_patterns.jl
concretization_patterns = ["const GLOBAL_CODE_STORE = Dict()"]

# logs toplevel analysis into toplevel.txt with debug logging level
toplevel_logger = """IOContext(open("toplevel.txt", "w"), :JET_LOGGER_LEVEL => 1)"""

Note

JET configurations specified as keyword arguments have precedence over those specified via a configuration file.