JET.jl Pluggable Analysis Framework

JET.jl Pluggable Analysis Framework

Interfaces

Examples

Warning

The APIs described in this page is very experimental and subject to changes. And this documentation is also very WIP.

Interfaces

JET.AbstractAnalyzer — Type

abstract type AbstractAnalyzer <: AbstractInterpreter end

When T implements this interface (i.e. T <: AbstractAnalyzer), T is expected to implement:

T(; jetconfigs...) -> T: constructs new analyzer given JET configurations; AnalyzerState for this analyzer should be constructed using these configurations

AnalyzerState(analyzer::T) -> AnalyzerState: returns AnalyzerState instance, usually it's kept within T itself

AbstractAnalyzer(analyzer::T, state::AnalyzerState) -> T: constructs new analyzer given the previous analyzer and analysis state

ReportPass(analyzer::T) -> ReportPass: returns the report pass of T

For example, JET.jl defines JETAnalyzer <: AbstractAnalyzer as the following (modified a bit for the sake of simplicity):

# the default abstract interpreter for JET.jl
struct JETAnalyzer{RP<:ReportPass} <: AbstractAnalyzer
    report_pass::RP
    state::AnalyzerState
end

# AbstractAnalyzer API requirements

function JETAnalyzer(;
    report_pass::Union{Nothing,T} = nothing,
    mode::Symbol                  = :basic,
    jetconfigs...) where {T<:ReportPass}
    if isnothing(report_pass)
        # if `report_pass` isn't passed explicitly, here we configure it according to `mode`
        report_pass = mode === :basic ? BasicPass() :
                      mode === :sound ? SoundPass() :
                      throw(ArgumentError("`mode` configuration should be either of `:basic` or `:sound`"))
    end
    return JETAnalyzer(report_pass,
                       AnalyzerState(; jetconfigs...),
                       )
end
AnalyzerState(analyzer::JETAnalyzer)                          = analyzer.state
AbstractAnalyzer(analyzer::JETAnalyzer, state::AnalyzerState) = JETAnalyzer(ReportPass(analyzer), state)
ReportPass(analyzer::JETAnalyzer)                             = analyzer.report_pass

Missing docstring.

Missing docstring for AnalyzerState. Check Documenter's build log for details.

JET.InferenceErrorReport — Type

InferenceErrorReport

An interface type of error reports that JET collects by abstract interpration. If T implements this interface, the following requirements should be satisfied:

Required fields
T should have the following fields, which explains where and why this error is reported:

vst::VirtualStackTrace: a virtual stack trace of the error

msg::String: explains why this error is reported

sig::Vector{Any}: a signature of the error point

Note that T can still have additional, specific fields.

A constructor interface to create T from abstraction interpretation
T<:InferenceErrorReport has the default constructor

T(::AbstractAnalyzer, sv::InferenceState, spec_args...)

which works when T is reported when sv's program counter (sv.currpc) points to that of statement where the error may happen. If so T just needs to overload

JET.get_msg(::Type{T}, ::AbstractAnalyzer, ::InferenceState, spec_args...) -> msg::String

to provide the message that describes why this error is reported (otherwise the senseless default message will be used).

If T is reported when sv's program counter (sv.currpc) may not point to the error location or even sv::InferenceState isn't available, T can implement its own constructor method.

A contructor interface to create T from the global report cache
In order to be cached and restored from JET_REPORT_CACHE, T must implement the following interfaces:

JET.get_spec_args(::T) -> Tuple{...}: returns fields that are specific to T, which is internally used by the caching logic

T(vst::VirtualStackTrace, msg::String, sig::Vector{Any} spec_args::Tuple{...}) -> T: constructor to create T from the cache, which should expand spec_args into each specific field

To satisfy these requirements manually will be very tedious. JET internally uses @reportdef utility macro, which takes the struct definition of InferenceErrorReport and automatically defines the struct itself and the cache interfaces.

See also: VirtualStackTrace, VirtualFrame

Missing docstring.

Missing docstring for get_msg. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_spec_args. Check Documenter's build log for details.

Missing docstring.

Missing docstring for @reportdef. Check Documenter's build log for details.

JET.ReportPass — Type

ReportPass

An interface type for report passes of JET's analysis.

Missing docstring.

Missing docstring for report_pass!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for report!. Check Documenter's build log for details.

Examples

Dispatch Analysis

Motivation

Implementation

Usages

"Unstable API" Analysis

Motivation

Implementation

Usages