JET Analyzer Framework

JET Analyzer Framework

Interfaces

Examples

JET offers an infrastructure to implement a "plugin" code analyzer. Actually, JET's default error analyzer that explained in the usage section is one specific instance of such a pluggin analyzer built on top of the framework.

In this documentation we will try to elaborate the framework APIs and showcase example analyzers.

Warning

The APIs described in this page is very experimental and subject to changes. And this documentation is also very WIP.

Interfaces

JET.AbstractAnalyzer — Type

abstract type AbstractAnalyzer <: AbstractInterpreter end

An interface type of analyzers that are built on top of JET's analyzer framework.

When a new type NewAnalyzer implements the AbstractAnalyzer interface, it should be declared as subtype of AbstractAnalyzer, and is expected to the following interfaces:

NewAnalyzer(; jetconfigs...) -> NewAnalyzer:
Constructs new analyzer given JET configurations passed as jetconfigs.

AnalyzerState(analyzer::NewAnalyzer) -> AnalyzerState:
Returns the AnalyzerState for analyzer::NewAnalyzer.

AbstractAnalyzer(analyzer::NewAnalyzer, state::AnalyzerState) -> NewAnalyzer:
Constructs an new NewAnalyzer instance in the middle of JET's top-level analysis or abstract interpretation based analysis, given the previous analyzer::NewAnalyzer and state::AnalyzerState.

ReportPass(analyzer::NewAnalyzer) -> ReportPass:
Returns ReportPass used for analyzer::NewAnalyzer.

See also AnalyzerState and ReportPass.

Example

JET.jl defines its default error analyzer JETAnalyzer <: AbstractAnalyzer as the following (modified a bit for the sake of simplicity):

# the default error analyzer for JET.jl
struct JETAnalyzer{RP<:ReportPass} <: AbstractAnalyzer
    report_pass::RP
    state::AnalyzerState
end

# AbstractAnalyzer API requirements

function JETAnalyzer(;
    report_pass::ReportPass = BasicPass(),
    jetconfigs...)
    return JETAnalyzer(report_pass,
                       AnalyzerState(; jetconfigs...),
                       )
end
AnalyzerState(analyzer::JETAnalyzer) =
    return analyzer.state
AbstractAnalyzer(analyzer::JETAnalyzer, state::AnalyzerState) =
    return JETAnalyzer(ReportPass(analyzer), state)
ReportPass(analyzer::JETAnalyzer) =
    return analyzer.report_pass

JET.AnalyzerState — Type

mutable struct AnalyzerState
    ...
end

The mutable object that holds various states that are consumed by all AbstractAnalyzers.

AnalyzerState(analyzer::AbstractAnalyzer) -> AnalyzerState

If NewAnalyzer implements the AbstractAnalyzer interface, NewAnalyzer should implement this AnalyzerState(analyzer::NewAnalyzer) -> AnalyzerState interface.

A new AnalyzerState is supposed to be constructed using JET configurations passed as keyword arguments jetconfigs of the NewAnalyzer(; jetconfigs...) constructor, and the constructed AnalyzerState is usually kept within NewAnalyzer itself:

function NewAnalyzer(; jetconfigs...)
    ...
    state = AnalyzerState(; jetconfigs...)
    return NewAnalyzer(..., state)
end
AnalyzerState(analyzer::NewAnalyzer) = analyzer.state

JET.ReportPass — Type

abstract type ReportPass end

An interface type that represents AbstractAnalyzer's report pass. analyzer::AbstractAnalyzer injects report passes using the (::ReportPass)(::Type{InferenceErrorReport}, ::AbstractAnalyzer, state, ...) interface, which provides a flexible and efficient layer to configure the analysis done by AbstractAnalyzer.

ReportPass(analyzer::AbstractAnalyzer) -> ReportPass

If NewAnalyzer implements the AbstractAnalyzer interface, NewAnalyzer should implement this ReportPass(analyzer::NewAnalyzer) -> ReportPass interface.

ReportPass allows NewAnalyzer to provide a very flexible configuration layer for NewAnalyzer's analysis; an user can define their own ReportPass to control how NewAnalyzer collects report errors while still using the analysis routine implemented by NewAnalyzer.

Example

For example, JETAnalyzer accepts a custom ReportPass passed as JET configurations (see the example documentation of AbstractAnalyzer for details). And we can setup a custom report pass IgnoreAllExceptGlobalUndefVar, that ignores all the reports that are otherwise collected by JETAnalyzer except GlobalUndefVarErrorReport:

# custom report pass that ignores all the reports except `GlobalUndefVarErrorReport`
struct IgnoreAllExceptGlobalUndefVar <: ReportPass end

# ignores all the reports analyzed by `JETAnalyzer`
(::IgnoreAllExceptGlobalUndefVar)(::Type{<:InferenceErrorReport}, @nospecialize(_...)) = return

# bypass to `BasicPass` to collect `GlobalUndefVarErrorReport`
function (::IgnoreAllExceptGlobalUndefVar)(::Type{GlobalUndefVarErrorReport}, @nospecialize(args...))
    BasicPass()(GlobalUndefVarErrorReport, args...)
end

no_method_error()    = 1 + "1"
undef_global_error() = undefvar
report_call(; report_pass=IgnoreAllExceptGlobalUndefVar()) do
    if rand(Bool)
        return no_method_error()    # "no matching method found" error report won't be reported here
    else
        return undef_global_error() # "variable undefvar is not defined" error report will be reported
    end
end

JET.InferenceErrorReport — Type

InferenceErrorReport

An interface type of error reports that JET collects by abstract interpration. If T implements this interface, the following requirements should be satisfied:

Required fields
T should have the following fields, which explains where and why this error is reported:

vst::VirtualStackTrace: a virtual stack trace of the error

msg::String: explains why this error is reported

sig::Vector{Any}: a signature of the error point

Note that T can still have additional, specific fields.

A constructor interface to create T from abstraction interpretation
T<:InferenceErrorReport has the default constructor

T(::AbstractAnalyzer, sv::InferenceState, spec_args...)

which works when T is reported when sv's program counter (sv.currpc) points to that of statement where the error may happen. If so T just needs to overload

JET.get_msg(::Type{T}, ::AbstractAnalyzer, ::InferenceState, spec_args...) -> msg::String

to provide the message that describes why this error is reported (otherwise the senseless default message will be used).

If T is reported when sv's program counter (sv.currpc) may not point to the error location or even sv::InferenceState isn't available, T can implement its own constructor method.

A contructor interface to create T from the global report cache
In order to be cached and restored from JET_CACHE, T must implement the following interfaces:

JET.get_spec_args(::T) -> Tuple{...}: returns fields that are specific to T, which is internally used by the caching logic

T(vst::VirtualStackTrace, msg::String, sig::Vector{Any} spec_args::Tuple{...}) -> T: constructor to create T from the cache, which should expand spec_args into each specific field

To satisfy these requirements manually will be very tedious. JET internally uses @reportdef utility macro, which takes the struct definition of InferenceErrorReport and automatically defines the struct itself and the cache interfaces.

See also: VirtualStackTrace, VirtualFrame

Missing docstring.

Missing docstring for get_msg. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_spec_args. Check Documenter's build log for details.

Missing docstring.

Missing docstring for @reportdef. Check Documenter's build log for details.

JET.add_new_report! — Function

add_new_report!(result::InferenceResult, report::InferenceErrorReport)

Adds new report::InferenceErrorReport to result::InferenceResult. result.src is supposed to be JETResult.

Examples

Dispatch Analysis

Motivation

Implementation

Usages

"Unstable API" Analysis

Motivation

Implementation

Usages